+++ emantomo/tests/test_protocols_emantomo.py
@@ -137,115 +137,6 @@ class TestEmanTomoBase(TestEmanBase):
         cls.smallTomogram = cls.dataset.getFile('coremask_normcorona.mrc')
 
 
-class TestEmanTomoExtraction(TestEmanTomoBase):
-    """This class check if the protocol to extract subtomograms
-    in Eman works properly.
-    """
-
-    @classmethod
-    def setUpClass(cls):
-        setupTestProject(cls)
-        TestEmanTomoBase.setData()
-
-    def _runTomoExtraction(self, tomoSource=0, doInvert=False, doNormalize=False, boxSize=32, downFactor=1):
-        protImportTomogram = self.newProtocol(tomo.protocols.ProtImportTomograms,
-                                              filesPath=self.tomogram,
-                                              samplingRate=5)
-
-        self.launchProtocol(protImportTomogram)
-
-        protImportCoordinates3d = self.newProtocol(tomo.protocols.ProtImportCoordinates3D,
-                                                   auto=IMPORT_FROM_EMAN,
-                                                   filesPath=self.coords3D,
-                                                   importTomograms=protImportTomogram.outputTomograms,
-                                                   filesPattern='', boxSize=32,
-                                                   samplingRate=5)
-
-        self.launchProtocol(protImportCoordinates3d)
-        self.assertSetSize(protImportTomogram.outputTomograms, 1,
-                           "There was a problem with tomogram output")
-        self.assertSetSize(protImportCoordinates3d.outputCoordinates, 5,
-                           "There was a problem with coordinates 3d output")
-
-        if tomoSource == 1:
-            protTomoExtraction = self.newProtocol(EmanProtTomoExtraction,
-                                                  inputTomograms=protImportTomogram.outputTomograms,
-                                                  inputCoordinates=protImportCoordinates3d.outputCoordinates,
-                                                  tomoSource=tomoSource,
-                                                  doInvert=doInvert,
-                                                  doNormalize=doNormalize,
-                                                  boxSize=boxSize,
-                                                  downFactor=downFactor)
-        else:
-            protTomoExtraction = self.newProtocol(EmanProtTomoExtraction,
-                                                  inputCoordinates=protImportCoordinates3d.outputCoordinates,
-                                                  tomoSource=tomoSource,
-                                                  doInvert=doInvert,
-                                                  doNormalize=doNormalize,
-                                                  boxSize=boxSize,
-                                                  downFactor=downFactor)
-        self.launchProtocol(protTomoExtraction)
-        self.assertSetSize(getattr(protTomoExtraction, OutputExtraction.subtomograms.name), 5,
-                           "There was a problem with SetOfSubtomogram output")
-        return protTomoExtraction
-
-    def test_extractParticlesSameAsPicking(self):
-        protTomoExtraction = self._runTomoExtraction()
-        output = getattr(protTomoExtraction, OutputExtraction.subtomograms.name, None)
-        self.assertTrue(output)
-        self.assertTrue(output.hasCoordinates3D())
-        self.assertTrue(output.getCoordinates3D().getObjValue())
-        return protTomoExtraction
-
-    def test_extractParticlesOther(self):
-        protTomoExtraction = self._runTomoExtraction(tomoSource=1)
-        output = getattr(protTomoExtraction, OutputExtraction.subtomograms.name, None)
-        self.assertTrue(output)
-        self.assertTrue(output.hasCoordinates3D())
-        self.assertTrue(output.getCoordinates3D().getObjValue())
-        return protTomoExtraction
-
-    def test_extractParticlesWithDoInvert(self):
-        protTomoExtraction = self._runTomoExtraction(doInvert=True)
-        output = getattr(protTomoExtraction, OutputExtraction.subtomograms.name, None)
-        self.assertTrue(output)
-        self.assertTrue(output.hasCoordinates3D())
-        self.assertTrue(output.getCoordinates3D().getObjValue())
-        return protTomoExtraction
-
-    def test_extractParticlesWithDoNormalize(self):
-        protTomoExtraction = self._runTomoExtraction(doNormalize=True)
-        output = getattr(protTomoExtraction, OutputExtraction.subtomograms.name, None)
-        self.assertTrue(output)
-        self.assertTrue(output.hasCoordinates3D())
-        self.assertTrue(output.getCoordinates3D().getObjValue())
-        return protTomoExtraction
-
-    def test_extractParticlesModifiedDownFactor(self):
-        protTomoExtraction = self._runTomoExtraction(downFactor=2)
-        output = getattr(protTomoExtraction, OutputExtraction.subtomograms.name, None)
-        self.assertTrue(output)
-        self.assertTrue(output.hasCoordinates3D())
-        self.assertTrue(output.getCoordinates3D().getObjValue())
-        return protTomoExtraction
-
-    def test_extractParticlesModifiedBoxSize(self):
-        protTomoExtraction = self._runTomoExtraction(boxSize=64)
-        output = getattr(protTomoExtraction, OutputExtraction.subtomograms.name, None)
-        self.assertTrue(output)
-        self.assertTrue(output.hasCoordinates3D())
-        self.assertTrue(output.getCoordinates3D().getObjValue())
-        return protTomoExtraction
-
-    def test_extractParticlesWithAllOptions(self):
-        protTomoExtraction = self._runTomoExtraction(boxSize=64, downFactor=2, doNormalize=True, doInvert=True)
-        output = getattr(protTomoExtraction, OutputExtraction.subtomograms.name, None)
-        self.assertTrue(output)
-        self.assertTrue(output.hasCoordinates3D())
-        self.assertTrue(output.getCoordinates3D().getObjValue())
-        return protTomoExtraction
-
-
 class TestEmanTomoInitialModel(TestEmanTomoBase):
     """This class check if the protocol to extract particles
     in Relion works properly.
@@ -394,157 +285,6 @@ class TestEmanTomoInitialModel(TestEmanTomoBase):
 
         return protInitialModel
 
-
-class TestEmanTomoSubtomogramRefinement(TestEmanTomoBase):
-    """This class check if the protocol Subtomogram refinement works properly.
-    """
-
-    @classmethod
-    def setUpClass(cls):
-        setupTestProject(cls)
-        TestEmanTomoBase.setData()
-
-    def _runPreviousProtocols(self):
-        protImportTomogram = self.newProtocol(tomo.protocols.ProtImportTomograms,
-                                              filesPath=self.tomogram,
-                                              samplingRate=5)
-
-        self.launchProtocol(protImportTomogram)
-
-        coords = pwutils.removeBaseExt(self.coords3D)
-        coords = protImportTomogram._getExtraPath(coords + '.txt')
-        pwutils.createAbsLink(self.coords3D_Large, coords)
-        protImportCoordinates3d = self.newProtocol(tomo.protocols.ProtImportCoordinates3D,
-                                                   auto=IMPORT_FROM_EMAN,
-                                                   filesPath=coords,
-                                                   importTomograms=protImportTomogram.outputTomograms,
-                                                   filesPattern='', boxSize=32,
-                                                   samplingRate=5)
-
-        self.launchProtocol(protImportCoordinates3d)
-        self.assertIsNotNone(protImportTomogram.outputTomograms,
-                             "There was a problem with tomogram output")
-        self.assertIsNotNone(protImportCoordinates3d.outputCoordinates,
-                             "There was a problem with coordinates 3d output")
-        doInvert = False
-        doNormalize = False
-        boxSize = 32
-        protTomoExtraction = self.newProtocol(EmanProtTomoExtraction,
-                                              inputTomograms=protImportTomogram.outputTomograms,
-                                              inputCoordinates=protImportCoordinates3d.outputCoordinates,
-                                              tomoSource=0,
-                                              doInvert=doInvert,
-                                              doNormalize=doNormalize,
-                                              boxSize=boxSize)
-
-        self.launchProtocol(protTomoExtraction)
-        self.assertIsNotNone(getattr(protTomoExtraction, OutputExtraction.subtomograms.name),
-                             "There was a problem with SetOfSubtomogram output")
-
-        return protTomoExtraction
-
-    def _performFinalValidation(self, protTomoSubtomogramRefinement):
-        outputSetOfSubTomograms = getattr(protTomoSubtomogramRefinement, EmanTomoRefinementOutputs.subtomograms.name)
-        averageSubTomogram = getattr(protTomoSubtomogramRefinement, EmanTomoRefinementOutputs.subtomogramAverage.name)
-
-        self.assertTrue(os.path.exists(averageSubTomogram.getFileName()))
-
-        # Check average is a mrc file
-        self.assertTrue(averageSubTomogram.getFileName().endswith(".mrc") )
-
-        self.assertEqual(averageSubTomogram.getSamplingRate(), 5.0)
-
-        self.assertEqual(outputSetOfSubTomograms.getDimensions(), (32, 32, 32))
-        self.assertEqual(outputSetOfSubTomograms.getSize(), 15)
-        self.assertEqual(outputSetOfSubTomograms.getCoordinates3D().getSize(), 15)
-
-        for subTomogram in outputSetOfSubTomograms:
-            self.assertEqual(subTomogram.getSamplingRate(), 5)
-            self.assertTrue(hasattr(subTomogram, EMAN_COVERAGE))
-            self.assertTrue(hasattr(subTomogram, EMAN_SCORE))
-            matrix = subTomogram.getTransform(convention=TR_EMAN).getMatrix()
-            self.assertEqual(matrix.shape, (4, 4))
-
-    def _runTomoSubtomogramRefinementWithSubtomo(self, niter=2, mass=500.0, threads=1, pkeep=1, goldstandard=-1,
-                                                 goldcontinue=False, sym="c1", localfilter=False, maxtilt=90.0):
-        protTomoExtraction = self._runPreviousProtocols()
-        protTomoRefinement = self.newProtocol(EmanProtTomoRefinement,
-                                              inputSetOfSubTomogram=getattr(protTomoExtraction, OutputExtraction.subtomograms.name),
-                                              inputRef=protTomoExtraction,
-                                              niter=niter,
-                                              mass=mass,
-                                              threads=threads,
-                                              pkeep=pkeep,
-                                              goldstandard=goldstandard,
-                                              goldcontinue=goldcontinue,
-                                              sym=sym,
-                                              localfilter=localfilter,
-                                              maxtilt=maxtilt)
-        protTomoRefinement.inputRef.setExtended("%s.1" % OutputExtraction.subtomograms.name)
-
-        self.launchProtocol(protTomoRefinement)
-
-        self.assertIsNotNone(getattr(protTomoRefinement, EmanTomoRefinementOutputs.subtomogramAverage.name),
-                             "There was a problem with subTomograms output")
-        self.assertSetSize(getattr(protTomoRefinement, EmanTomoRefinementOutputs.subtomograms.name), None,
-                             "There was a problem with particles output")
-
-        return protTomoRefinement
-
-    def _runTomoSubtomogramRefinementWithVolume(self, niter=2, mass=500.0, threads=1, pkeep=1, goldstandard=-1,
-                                                goldcontinue=False, sym="c1", localfilter=False, maxtilt=90.0):
-        protTomoExtraction = self._runPreviousProtocols()
-
-        particles = getattr(protTomoExtraction, OutputExtraction.subtomograms.name)
-
-        self.dataset = DataSet.getDataSet('eman')
-        self.vol = self.dataset.getFile('volume')
-        self.protImportVol = self.runImportVolumes(self.vol, 3.5)
-
-        self.assertIsNotNone(self.protImportVol.outputVolume,
-                             "There was a problem with SetOfSubtomogram output")
-
-        protTomoRefinement = self.newProtocol(EmanProtTomoRefinement,
-                                              inputSetOfSubTomogram=particles,
-                                              inputRef=self.protImportVol.outputVolume,
-                                              niter=niter,
-                                              mass=mass,
-                                              threads=threads,
-                                              pkeep=pkeep,
-                                              goldstandard=goldstandard,
-                                              goldcontinue=goldcontinue,
-                                              sym=sym,
-                                              localfilter=localfilter,
-                                              maxtilt=maxtilt)
-
-        self.launchProtocol(protTomoRefinement)
-
-        average = getattr(protTomoRefinement, EmanTomoRefinementOutputs.subtomogramAverage.name)
-        self.assertIsNotNone(average,  "There was a problem with average output")
-        self.assertTrue(os.path.exists(average.getFileName()), "Average %s does not exists" % average.getFileName())
-
-        self.assertTrue(average.hasHalfMaps(), "Halves not registered")
-
-        half1, half2 = average.getHalfMaps().split(',')
-        self.assertTrue(os.path.exists(half1), msg="Average 1st half %s does not exists" % half1)
-        self.assertTrue(os.path.exists(half2), msg="Average 2nd half %s does not exists" % half2)
-
-        self.assertSetSize(getattr(protTomoRefinement, EmanTomoRefinementOutputs.subtomograms.name), None,
-                             "There was a problem with particles output")
-
-        return protTomoRefinement
-
-    def test_defaultSubTomogramRefinementWithSubTomo(self):
-        protTomoSubtomogramRefinement = self._runTomoSubtomogramRefinementWithSubtomo()
-        self._performFinalValidation(protTomoSubtomogramRefinement)
-
-        return protTomoSubtomogramRefinement
-
-    def test_defaultSubTomogramRefinementWithVolume(self):
-        protTomoSubtomogramRefinement = self._runTomoSubtomogramRefinementWithVolume()
-        self._performFinalValidation(protTomoSubtomogramRefinement)
-
-        return protTomoSubtomogramRefinement
 
 
 class TestEmanTomoTempMatch(TestEmanTomoBase):
